Fix: storybook fetching error - extract view component
